+++ docs/conf.py
@@ -183,6 +183,7 @@ intersphinx_mapping = {
     "venv": ("https://virtualenv.pypa.io/en/stable/", None),
     "borg": ("https://borgbackup.readthedocs.io/en/stable/", None),
     "pip": ("https://pip.pypa.io/en/stable/", None),
+    "compressor": ("https://django-compressor.readthedocs.io/en/stable/", None),
 }
 
 # Ignore missing targets for the http:obj <type>, it's how we declare the types