+++ weblate/addons/git.py
@@ -202,7 +202,7 @@ class GitSquashAddon(BaseAddon):
             repository.execute(["checkout", repository.branch])
             repository.delete_branch(tmp)
 
-    def post_commit(self, component, translation=None):
+    def post_commit(self, component):
         repository = component.repository
         with repository.lock:
             if component.repo_needs_merge() and not component.update_branch(