+++ weblate/static/editor/full.js
@@ -140,77 +140,56 @@
             Cookies.set('translate-tab', $(this).attr('href'), { path: '/', expires: 365 });
         });
 
-        /* Machine translation */
-        this.isTMLoaded = false;
-        this.$editor.on('show.bs.tab', '[data-load="mt"]', () => {
-            if (this.isMTLoaded) {
+        /* Machinery */
+        this.isMachineryLoaded = false;
+        this.$editor.on('show.bs.tab', '[data-load="machinery"]', () => {
+            if (this.isMachineryLoaded) {
                 return;
             }
-            this.isMTLoaded = true;
-            this.initMachineTranslation();
-        });
-
-        /* Translation memory */
-        this.isMTLoaded = false;
-        this.$editor.on('show.bs.tab', '[data-load="memory"]', () => {
-            if (this.isTMLoaded) {
-                return;
-            }
-            this.isTMLoaded = true;
-            this.initTranslationMemory();
+            this.initMachinery();
         });
     };
 
-    FullEditor.prototype.initMachineTranslation = function () {
-        increaseLoading('mt');
-        $.ajax({
-            url: $('#js-mt-services').attr('href'),
-            success: (servicesList) => {
-                decreaseLoading('mt');
-                servicesList.forEach((serviceName) => {
-                    increaseLoading('mt');
-                    this.fetchMachinery(serviceName, 'mt');
-                });
-            },
-            error: this.processMachineryError,
-            dataType: 'json'
-        });
-    };
+    FullEditor.prototype.initMachinery = function () {
+        this.isMachineryLoaded = true;
+        this.machinery = new Machinery();
 
-    FullEditor.prototype.initTranslationMemory = function () {
-        increaseLoading('memory');
-        this.fetchMachinery(TM_SERVICE_NAME, 'memory');
+        $("#js-translate").data("services").forEach((serviceName) => {
+              increaseLoading('machinery');
+              this.fetchMachinery(serviceName);
+        });
 
         this.$editor.on('submit', '#memory-search', (e) => {
             var $form = $(e.currentTarget);
 
-            increaseLoading('memory');
-            $('#memory-translations').empty();
+            increaseLoading('machinery');
+            this.machinery.setState({translations: []});
+            $('#machinery-translations').empty();
             $.ajax({
                 type: 'POST',
                 url: $form.attr('action'),
                 data: $form.serialize(),
                 dataType: 'json',
                 success: (data) => {
-                    this.processMachineryResults(data, 'memory');
+                    this.processMachineryResults(data);
                 },
                 error: (jqXHR, textStatus, errorThrown) => {
-                    this.processMachineryError(jqXHR, textStatus, errorThrown, 'memory');
+                    this.processMachineryError(jqXHR, textStatus, errorThrown);
                 },
             });
             return false;
         });
     };
 
-    FullEditor.prototype.fetchMachinery = function (serviceName, serviceType) {
+    FullEditor.prototype.fetchMachinery = function (serviceName) {
         $.ajax({
             type: 'POST',
             url: $('#js-translate').attr('href').replace('__service__', serviceName),
             success: (data) => {
-                this.processMachineryResults(data, serviceType);
+                this.processMachineryResults(data);
             },
             error: (jqXHR, textStatus, errorThrown) => {
-                this.processMachineryError(jqXHR, textStatus, errorThrown, serviceType);
+                this.processMachineryError(jqXHR, textStatus, errorThrown);
             },
             dataType: 'json',
             data: {
@@ -219,15 +198,15 @@
         });
     };
 
-    FullEditor.prototype.processMachineryError = function (jqXHR, textStatus, errorThrown, scope) {
-        decreaseLoading(scope);
+    FullEditor.prototype.processMachineryError = function (jqXHR, textStatus, errorThrown) {
+        decreaseLoading('machinery');
         if (jqXHR.state() !== 'rejected') {
             addAlert(gettext('The request for machine translation has failed:') + ' ' + textStatus + ': ' + errorThrown);
         }
     };
 
-    FullEditor.prototype.processMachineryResults = function (data, scope) {
-        decreaseLoading(scope);
+    FullEditor.prototype.processMachineryResults = function (data) {
+        decreaseLoading('machinery');
         if (data.responseStatus !== 200) {
             var msg = interpolate(
                 gettext('The request for machine translation using %s has failed:'),
@@ -238,79 +217,12 @@
             return;
         }
 
-        data.translations.forEach(function (el, idx) {
-            var newRow = $('<tr/>').data('raw', el);
-            var done = false;
-            var $translations = $('#' + scope + '-translations');
-            var service;
-
-            newRow.append($('<td/>').attr('class', 'target mt-text').attr('lang', data.lang).attr('dir', data.dir).text(el.text));
-            newRow.append($('<td/>').attr('class', 'mt-text').text(el.source));
-            if (scope === "mt") {
-                service = $('<td/>').text(el.service);
-                if (typeof el.origin !== 'undefined') {
-                    service.append(' (');
-                    var origin;
-                    if (typeof el.origin_detail !== 'undefined') {
-                        origin = $('<abbr/>').text(el.origin).attr('title', el.origin_detail);
-                    } else if (typeof el.origin_url !== 'undefined') {
-                        origin = $('<a/>').text(el.origin).attr('href', el.origin_url);
-                    } else {
-                        origin = el.origin;
-                    }
-                    service.append(origin);
-                    service.append(')');
-                }
-            } else {
-                service = $('<td/>').text(el.origin);
-            }
-            newRow.append(service);
-            /* Quality score as bar with the text */
-            newRow.append($(
-                '<td>' +
-                '<div class="progress" title="' + el.quality + ' / 100">' +
-                '<div class="progress-bar ' +
-                (el.quality >= 70 ? 'progress-bar-success' : el.quality >= 50 ? 'progress-bar-warning' : 'progress-bar-danger') + '"' +
-                ' role="progressbar" aria-valuenow="' + el.quality + '"' +
-                ' aria-valuemin="0" aria-valuemax="100" style="width: ' + el.quality + '%;"></div>' +
-                '</div>' +
-                '</td>'
-            ));
-            /* Translators: Verb for copy operation */
-            newRow.append($(
-                '<td>' +
-                '<a class="js-copy-machinery btn btn-warning">' +
-                gettext('Copy') +
-                '<span class="mt-number text-info"></span>' +
-                '</a>' +
-                '</td>' +
-                '<td>' +
-                '<a class="js-copy-save-machinery btn btn-primary">' +
-                gettext('Copy and save') +
-                '</a>' +
-                '</td>'
-            ));
-            $translations.children('tr').each(function (idx) {
-                var $this = $(this);
-                var base = $this.data('raw');
-                if (base.text == el.text && base.source == el.source) {
-                    // Add origin to current ones
-                    var current = $this.children('td:nth-child(3)');
-                    current.append($("<br/>"));
-                    current.append(service.html());
-                    done = true;
-                    return false;
-                } else if (base.quality <= el.quality) {
-                    // Insert match before lower quality one
-                    $this.before(newRow);
-                    done = true;
-                    return false;
-                }
-            });
-            if (!done) {
-                $translations.append(newRow);
-            }
-        });
+        this.machinery.setState({
+            translations: [...this.machinery.state.translations, ...data.translations],
+            lang: data.lang,
+            dir: data.dir,
+        })
+        this.machinery.render(data.translations);
 
         // Cancel out browser's `meta+m` and let Mousetrap handle the rest
         document.addEventListener('keydown', function (e) {
@@ -321,7 +233,7 @@
             }
         });
 
-        var $translationRows = $('#' + scope + '-translations').children('tr');
+        var $translationRows = $('#machinery-translations').children('tr');
         $translationRows.each(function (idx) {
             if (idx < 10) {
                 var key = WLT.Utils.getNumericKey(idx);
@@ -332,7 +244,7 @@
                 } else {
                     title = interpolate(gettext('Ctrl+M then %s'), [key]);
                 }
-                $(this).find('.mt-number').html(
+                $(this).find('.machinery-number').html(
                     ' <kbd title="' + title + '">' + key + '</kbd>'
                 );
                 Mousetrap.bindGlobal(
@@ -343,7 +255,7 @@
                     }
                 );
             } else {
-                $(this).find('.mt-number').html('');
+                $(this).find('.machinery-number').html('');
             }
         });
     };
@@ -490,6 +402,108 @@
     FullEditor.prototype.insertIntoTranslation = function (text) {
         this.$translationArea.insertAtCaret($.trim(text)).change();
     };
+
+    class Machinery {
+        constructor(initialState = {}) {
+            this.state = {
+                translations: [],
+                lang: null,
+                dir: null,
+            };
+        }
+
+        setState(newState) {
+            this.state = {...this.state, ...newState};
+        }
+
+        renderTranslation(el, service) {
+            var row = $('<tr/>').attr('class', 'js-copy-machinery').data('raw', el);
+            row.append($('<td/>').attr('class', 'target machinery-text').attr('lang', this.state.lang).attr('dir', this.state.dir).text(el.text));
+            row.append($('<td/>').attr('class', 'machinery-text').text(el.source));
+            row.append(service);
+
+            /* Quality score as bar with the text */
+            row.append($(
+                '<td>' +
+                '<div class="progress" title="' + el.quality + ' / 100">' +
+                '<div class="progress-bar ' +
+                (el.quality >= 70 ? 'progress-bar-success' : el.quality >= 50 ? 'progress-bar-warning' : 'progress-bar-danger') + '"' +
+                ' role="progressbar" aria-valuenow="' + el.quality + '"' +
+                ' aria-valuemin="0" aria-valuemax="100" style="width: ' + el.quality + '%;"></div>' +
+                '</div>' +
+                '</td>'
+            ));
+            /* Translators: Verb for copy operation */
+            row.append($(
+                '<td>' +
+                '<a class="js-copy-machinery btn btn-warning">' +
+                gettext('Copy') +
+                '<span class="mt-number text-info"></span>' +
+                '</a>' +
+                '</td>' +
+                '<td>' +
+                '<a class="js-copy-save-machinery btn btn-primary">' +
+                gettext('Copy and save') +
+                '</a>' +
+                '</td>'
+            ));
+
+            return row;
+        }
+
+        renderService(el) {
+            var service = $('<td/>').text(el.service);
+            if (typeof el.origin !== 'undefined') {
+                service.append(' (');
+                var origin;
+                if (typeof el.origin_detail !== 'undefined') {
+                    origin = $('<abbr/>').text(el.origin).attr('title', el.origin_detail);
+                } else if (typeof el.origin_url !== 'undefined') {
+                    origin = $('<a/>').text(el.origin).attr('href', el.origin_url);
+                } else {
+                    origin = el.origin;
+                }
+                service.append(origin);
+                service.append(')');
+            }
+            return service;
+        }
+
+        render() {
+            var $translations = $('#machinery-translations');
+            this.state.translations.forEach((translation) => {
+                var service = this.renderService(translation);
+                var insertBefore;
+                var done = false;
+
+                /* This is the merging and insert sort logic */
+                $translations.children('tr').each(function (idx) {
+                    var $this = $(this);
+                    var base = $this.data('raw');
+                    if (base.text == translation.text && base.source == translation.source) {
+                        // Add origin to current ones
+                        var current = $this.children('td:nth-child(3)');
+                        current.append($("<br/>"));
+                        current.append(service.html());
+                        done = true;
+                        return false;
+                    } else if (base.quality <= translation.quality) {
+                        // Insert match before lower quality one
+                        insertBefore = $this;
+                    }
+                });
+
+                if (! done) {
+                    var newRow = this.renderTranslation(translation, service);
+                    if (insertBefore) {
+                        insertBefore.before(newRow);
+                    } else {
+                        $translations.append(newRow);
+                    }
+                }
+          });
+        }
+    }
 
     document.addEventListener('DOMContentLoaded', function () {
         new FullEditor();