+++ docs/changes.rst
@@ -17,6 +17,8 @@ Not yet released.
 * Improved performance of bulk state changes.
 * Added compatibility redirects after project or component renaming.
 * Added notifications for strings approval, component locking and license change.
+* Added support for ModernMT.
+* Allow to avoid overwriting approved translations on file upload.
 
 Weblate 4.1.1
 -------------